# app/Http/Controllers/SuperStar/SuperStarAuth.php
<?php

namespace App\Http\Controllers\SuperStar;

use App\Http\Controllers\Controller;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Hash;
use Illuminate\Support\Facades\Validator;
use App\Models\Superstar;

class SuperStarAuth extends Controller
{
    /**
     * @OA\Post(
     *     path="/api/superstar/login",
     *     summary="Superstar Login",
     *     description="Login endpoint for superstars using email or username",
     *     tags={"Superstar Authentication"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"login","password"},
     *             @OA\Property(property="login", type="string", description="Email or username", example="superstar@example.com"),
     *             @OA\Property(property="password", type="string", format="password", example="password123")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Login successful",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Login successful"),
     *             @OA\Property(property="superstar", type="object"),
     *             @OA\Property(property="token", type="string")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Invalid credentials"
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     */
    public function login(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'login' => 'required|string', // Can be email or username
            'password' => 'required|string',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'message' => 'Validation failed',
                'errors' => $validator->errors()
            ], 422);
        }

        $loginField = $request->input('login');
        $password = $request->input('password');

        // Check if login field is email or username
        $loginType = filter_var($loginField, FILTER_VALIDATE_EMAIL) ? 'email' : 'username';

        // Find superstar by email or username
        $superstar = Superstar::where($loginType, $loginField)->first();

        if (!$superstar) {
            return response()->json([
                'message' => 'Invalid credentials'
            ], 401);
        }

        // Check role - if admin, return error
        if ($superstar->role === 'admin') {
            return response()->json([
                'message' => 'Please use the admin app for admin access'
            ], 403);
        }

        // Verify password
        if (!Hash::check($password, $superstar->password)) {
            return response()->json([
                'message' => 'Invalid credentials'
            ], 401);
        }

        // Revoke existing tokens if any
        $superstar->tokens()->delete();

        // Create new sanctum token
        $token = $superstar->createToken('superstar-auth-token')->plainTextToken;

        return response()->json([
            'message' => 'Login successful',
            'superstar' => [
                'id' => $superstar->id,
                'user_id' => $superstar->user_id,
                'display_name' => $superstar->display_name,
                'bio' => $superstar->bio,
                'price_per_hour' => $superstar->price_per_hour,
                'is_available' => $superstar->is_available,
                'rating' => $superstar->rating,
                'total_followers' => $superstar->total_followers,
                'status' => $superstar->status,
                'username' => $superstar->username,
                'email' => $superstar->email,
                'role' => $superstar->role,
                'created_at' => $superstar->created_at
            ],
            'token' => $token,
            'token_type' => 'Bearer'
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/superstar/logout",
     *     summary="Logout superstar",
     *     description="Logout the authenticated superstar and revoke token",
     *     tags={"Superstar Authentication"},
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Logout successful",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Logged out successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     )
     * )
     */
    public function logout(Request $request)
    {
        $request->user()->currentAccessToken()->delete();

        return response()->json([
            'message' => 'Logged out successfully'
        ]);
    }

    /**
     * @OA\Get(
     *     path="/api/superstar/me",
     *     summary="Get authenticated superstar",
     *     description="Get the authenticated superstar's profile information",
     *     tags={"Superstar Authentication"},
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Profile retrieved successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="superstar", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthorized"
     *     )
     * )
     */
    public function me(Request $request)
    {
        $superstar = $request->user();

        return response()->json([
            'superstar' => [
                'id' => $superstar->id,
                'user_id' => $superstar->user_id,
                'display_name' => $superstar->display_name,
                'bio' => $superstar->bio,
                'price_per_hour' => $superstar->price_per_hour,
                'is_available' => $superstar->is_available,
                'rating' => $superstar->rating,
                'total_followers' => $superstar->total_followers,
                'status' => $superstar->status,
                'username' => $superstar->username,
                'email' => $superstar->email,
                'role' => $superstar->role,
                'created_at' => $superstar->created_at,
                'updated_at' => $superstar->updated_at
            ]
        ]);
    }

    /**
     * @OA\Put(
     *     path="/api/superstar/profile",
     *     summary="Update superstar profile",
     *     description="Update the authenticated superstar's profile information",
     *     tags={"Superstar Profile"},
     *     security={{"sanctum":{}}},
     *     @OA\RequestBody(
     *         required=false,
     *         @OA\JsonContent(
     *             @OA\Property(property="display_name", type="string", example="John Doe"),
     *             @OA\Property(property="bio", type="string", example="Professional superstar"),
     *             @OA\Property(property="price_per_hour", type="number", format="float", example=50.00),
     *             @OA\Property(property="is_available", type="boolean", example=true),
     *             @OA\Property(property="username", type="string", example="johndoe"),
     *             @OA\Property(property="email", type="string", format="email", example="john@example.com")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Profile updated successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Profile updated successfully"),
     *             @OA\Property(property="superstar", type="object")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     */
    public function updateProfile(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'display_name' => 'sometimes|string|max:255',
            'bio' => 'sometimes|string|max:1000',
            'price_per_hour' => 'sometimes|numeric|min:0|max:9999.99',
            'is_available' => 'sometimes|boolean',
            'username' => 'sometimes|string|max:255|unique:superstars,username,' . $request->user()->id,
            'email' => 'sometimes|email|max:255|unique:superstars,email,' . $request->user()->id,
        ]);

        if ($validator->fails()) {
            return response()->json([
                'message' => 'Validation failed',
                'errors' => $validator->errors()
            ], 422);
        }

        $superstar = $request->user();
        
        $updateData = [];
        
        // Only update fields that are provided
        if ($request->has('display_name')) {
            $updateData['display_name'] = $request->display_name;
        }
        if ($request->has('bio')) {
            $updateData['bio'] = $request->bio;
        }
        if ($request->has('price_per_hour')) {
            $updateData['price_per_hour'] = $request->price_per_hour;
        }
        if ($request->has('is_available')) {
            $updateData['is_available'] = $request->boolean('is_available');
        }
        if ($request->has('username')) {
            $updateData['username'] = $request->username;
        }
        if ($request->has('email')) {
            $updateData['email'] = $request->email;
        }
        
        $superstar->update($updateData);
        
        return response()->json([
            'message' => 'Profile updated successfully',
            'superstar' => [
                'id' => $superstar->id,
                'user_id' => $superstar->user_id,
                'display_name' => $superstar->display_name,
                'bio' => $superstar->bio,
                'price_per_hour' => $superstar->price_per_hour,
                'is_available' => $superstar->is_available,
                'rating' => $superstar->rating,
                'total_followers' => $superstar->total_followers,
                'status' => $superstar->status,
                'username' => $superstar->username,
                'email' => $superstar->email,
                'role' => $superstar->role,
                'created_at' => $superstar->created_at,
                'updated_at' => $superstar->updated_at
            ]
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/superstar/change-password",
     *     summary="Change superstar password",
     *     description="Change the authenticated superstar's password",
     *     tags={"Superstar Authentication"},
     *     security={{"sanctum":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"current_password","password"},
     *             @OA\Property(property="current_password", type="string", format="password"),
     *             @OA\Property(property="password", type="string", format="password", minLength=8)
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Password changed successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Password changed successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error"
     *     )
     * )
     */
    public function changePassword(Request $request)
    {
        $validator = Validator::make($request->all(), [
            'current_password' => 'required|string',
            'new_password' => 'required|string|min:8|confirmed',
        ]);

        if ($validator->fails()) {
            return response()->json([
                'message' => 'Validation failed',
                'errors' => $validator->errors()
            ], 422);
        }

        $superstar = $request->user();
        
        // Verify current password
        if (!Hash::check($request->current_password, $superstar->password)) {
            return response()->json([
                'message' => 'Current password is incorrect'
            ], 422);
        }
        
        // Update password
        $superstar->update([
            'password' => Hash::make($request->new_password)
        ]);
        
        // Revoke all existing tokens (force logout from all devices)
        $superstar->tokens()->delete();
        
        return response()->json([
            'message' => 'Password changed successfully. Please login again.'
        ]);
    }
}