# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: Rashid Backend API Documentation
  description: Complete API Documentation for Rashid Backend Application
  version: 1.0.0
  contact:
    email: admin@rashid.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://127.0.0.1:8000
    description: Rashid Backend API Server
...
paths:
...
  /api/superstar/login:
    post:
      tags:
        - Superstar Authentication
      summary: Superstar Login
      description: Login endpoint for superstars using email or username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - login
                - password
              properties:
                login:
                  type: string
                  description: Email or username
                  example: superstar@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login successful
                  superstar:
                    type: object
                  token:
                    type: string
        '401':
          description: Invalid credentials
        '422':
          description: Validation error
  /api/superstar/logout:
    post:
      tags:
        - Superstar Authentication
      summary: Logout superstar
      description: Logout the authenticated superstar and revoke token
      security:
        - sanctum: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logged out successfully
        '401':
          description: Unauthorized
  /api/superstar/me:
    get:
      tags:
        - Superstar Authentication
      summary: Get authenticated superstar
      description: Get the authenticated superstar's profile information
      security:
        - sanctum: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  superstar:
                    type: object
        '401':
          description: Unauthorized

  # Superstar Profile
  /api/superstar/profile:
    put:
      tags:
        - Superstar Profile
      summary: Update superstar profile
      description: Update the authenticated superstar's profile information
      security:
        - sanctum: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  example: John Doe
                bio:
                  type: string
                  example: Professional superstar
                price_per_hour:
                  type: number
                  format: float
                  example: 50.00
                is_available:
                  type: boolean
                  example: true
                username:
                  type: string
                  example: johndoe
                email:
                  type: string
                  format: email
                  example: john@example.com
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Profile updated successfully
                  superstar:
                    type: object
        '422':
          description: Validation error
  /api/superstar/change-password:
    post:
      tags:
        - Superstar Authentication
      summary: Change superstar password
      description: Change the authenticated superstar's password
      security:
        - sanctum: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - current_password
                - password
              properties:
                current_password:
                  type: string
                  format: password
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password changed successfully
        '422':
          description: Validation error

  # Superstar Payments
...
components:
  securitySchemes:
    sanctum:
      type: apiKey
      in: header
      name: Authorization
      description: Enter token in format (Bearer <token>)